idk why anymore

+++ unit-9/football/Bouncing Ball Lab/src/Thirdball.java
@@ -4,7 +4,7 @@ public class Thirdball extends Ball {
 
     private int radius;
 
-    public Thirdball(double x, double y, Color c, int radius) {
+    public Thirdball(int x, int y, Color c, int radius) {
         super(x, y, c);
         this.radius = radius;
     }
